Abilities andhero classes initialized, added new ui for testing and hero selection later on. Some animations and new enemies

+++ app/src/main/scala/com/tume/scalarpg/model/item/EquipmentAffix.scala
@@ -8,14 +8,14 @@ import com.tume.scalarpg.model.property.{DamageRange, Healing, Damage}
   */
 sealed trait EquipmentAffix {
 
-  private val scalingStart = 10D
+  private val scalingStart = 10f
 
   var itemLevel = 1
-  var variation = 1D
+  var variation = 1f
 
   def weight(equipSlot: EquipSlot) : Int
   def tooltipLine : String
-  def amount: Double
+  def amount: Float
 
   def factor = 1 + amount
   def percent = (amount * 100).round + "%"
@@ -26,40 +26,40 @@ sealed trait EquipmentAffix {
   def modifyIncomingHealing(healing: Healing) : Healing = healing
 
   def modifyBaseDamage(damage: DamageRange) : DamageRange = damage
-  def modifyAbilityPower(power: Double) : Double = power
+  def modifyAbilityPower(power: Float) : Float = power
 
-  def modifyHealth(health: Double) : Double = health
-  def modifyMana(mana: Double) : Double = mana
-  def modifySpeed(speed: Double) : Double = speed
-  def modifyAttackSpeed(speed: Double) : Double = speed
-  def modifyMovementSpeed(speed: Double) : Double = speed
-  def modifyExperienceGain(xp: Double) : Double = xp
+  def modifyHealth(health: Float) : Float = health
+  def modifyMana(mana: Float) : Float = mana
+  def modifySpeed(speed: Float) : Float = speed
+  def modifyAttackSpeed(speed: Float) : Float = speed
+  def modifyMovementSpeed(speed: Float) : Float = speed
+  def modifyExperienceGain(xp: Float) : Float = xp
 
   def standardScaling = (scalingStart + (itemLevel * variation + (itemLevel * variation + 1)) / 2) / (scalingStart + 1)
   def logScaling = Math.log(scalingStart * itemLevel * variation) / Math.log(scalingStart + 1)
 }
 case class IncreasedHealthAffix() extends EquipmentAffix {
-  override def amount = 10D * standardScaling
+  override def amount = 10f * standardScaling
   override def weight(equipSlot: EquipSlot) : Int = equipSlot match {
     case Weapon => 10
     case Trinket => 100
     case _ => 100
   }
   override def tooltipLine = s"+$percent health"
-  override def modifyHealth(double: Double) = double * factor
+  override def modifyHealth(Float: Float) = Float * factor
 }
 case class IncreasedManaAffix() extends EquipmentAffix {
-  override def amount = 10D * standardScaling
+  override def amount = 10f * standardScaling
   override def weight(equipSlot: EquipSlot) : Int = equipSlot match {
     case Weapon => 10
     case Trinket => 150
     case _ => 50
   }
   override def tooltipLine = s"+$percent mana"
-  override def modifyMana(double: Double) = double * factor
+  override def modifyMana(Float: Float) = Float * factor
 }
 case class IncreasedBaseDamageAffix() extends EquipmentAffix {
-  override def amount = 10D * standardScaling
+  override def amount = 10f * standardScaling
   override def weight(equipSlot: EquipSlot) : Int = equipSlot match {
     case Weapon => 500
     case Trinket => 70
